App/requirements.txt
altair==5.4.0
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
narwhals==1.5.4
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.37.1
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
